+++ pypesto/hierarchical/base_solver.py
@@ -1,0 +1,42 @@
+from typing import Dict, List
+
+import numpy as np
+
+from .base_problem import InnerProblem
+
+
+class InnerSolver:
+    """Solver for an inner optimization problem."""
+
+    def initialize(self):
+        """
+        (Re-)initialize the solver.
+
+        Default: Do nothing.
+        """
+
+    def solve(
+        self,
+        problem: InnerProblem,
+        sim: List[np.ndarray],
+        sigma: List[np.ndarray],
+        scaled: bool,
+    ) -> Dict[str, float]:
+        """Solve the subproblem.
+
+        Parameters
+        ----------
+        problem:
+            The inner problem to solve.
+        sim:
+            List of model output matrices, as provided in AMICI's
+            ``ReturnData.y``. Same order as simulations in the
+            PEtab problem.
+        sigma:
+            List of sigma matrices from the model, as provided in AMICI's
+            ``ReturnData.sigmay``. Same order as simulations in the
+            PEtab problem.
+        scaled:
+            Whether to scale the results to the parameter scale specified in
+            ``problem``.
+        """
