+++ pypesto/visualize/spline_approximation.py
@@ -21,11 +21,11 @@ from ..result import Result
 try:
     import amici
 
-    from ..hierarchical.spline_approximation.calculator import (
-        SplineAmiciCalculator,
+    from ..hierarchical.semiquantitative.calculator import (
+        SemiquantitativeCalculator,
     )
-    from ..hierarchical.spline_approximation.solver import (
-        SplineInnerSolver,
+    from ..hierarchical.semiquantitative.solver import (
+        SemiquantitativeInnerSolver,
         _calculate_regularization_for_group,
         get_spline_mapped_simulations,
     )
@@ -107,7 +107,7 @@ def plot_splines_from_pypesto_result(
     for (
         calculator
     ) in pypesto_result.problem.objective.calculator.inner_calculators:
-        if isinstance(calculator, SplineAmiciCalculator):
+        if isinstance(calculator, SemiquantitativeCalculator):
             spline_calculator = calculator
             break
 
@@ -188,7 +188,11 @@ def plot_splines_from_inner_result(
         simulation = inner_problem.groups[group][CURRENT_SIMULATION]
 
         # For the simulation, get the spline bases
-        delta_c, spline_bases, n = SplineInnerSolver._rescale_spline_bases(
+        (
+            delta_c,
+            spline_bases,
+            n,
+        ) = SemiquantitativeInnerSolver._rescale_spline_bases(
             self=None,
             sim_all=simulation,
             N=len(inner_parameters),
@@ -353,7 +357,7 @@ def _add_spline_mapped_simulations_to_model_fit(
 
     spline_calculator = None
     for calculator in pypesto_problem.objective.calculator.inner_calculators:
-        if isinstance(calculator, SplineAmiciCalculator):
+        if isinstance(calculator, SemiquantitativeCalculator):
             spline_calculator = calculator
             break
 
@@ -384,7 +388,11 @@ def _add_spline_mapped_simulations_to_model_fit(
         simulation = inner_problem.groups[group][CURRENT_SIMULATION]
 
         # For the simulation, get the spline bases
-        delta_c, spline_bases, n = SplineInnerSolver._rescale_spline_bases(
+        (
+            delta_c,
+            spline_bases,
+            n,
+        ) = SemiquantitativeInnerSolver._rescale_spline_bases(
             self=None,
             sim_all=simulation,
             N=len(inner_parameters),
@@ -502,7 +510,7 @@ def _obtain_regularization_for_start(
     for (
         calculator
     ) in pypesto_result.problem.objective.calculator.inner_calculators:
-        if isinstance(calculator, SplineAmiciCalculator):
+        if isinstance(calculator, SemiquantitativeCalculator):
             spline_calculator = calculator
             break
 
@@ -525,7 +533,11 @@ def _obtain_regularization_for_start(
         simulation = inner_problem.groups[group][CURRENT_SIMULATION]
 
         # For the simulation, get the spline bases
-        delta_c, spline_bases, n = SplineInnerSolver._rescale_spline_bases(
+        (
+            delta_c,
+            spline_bases,
+            n,
+        ) = SemiquantitativeInnerSolver._rescale_spline_bases(
             self=None,
             sim_all=simulation,
             N=len(inner_parameters),
