+++ test/hierarchical/test_ordinal.py
@@ -24,14 +24,12 @@ from pypesto.C import (
     STANDARD,
     InnerParameterType,
 )
-from pypesto.hierarchical.optimal_scaling import (
+from pypesto.hierarchical.ordinal import (
     OptimalScalingInnerSolver,
     OptimalScalingProblem,
 )
-from pypesto.hierarchical.optimal_scaling.parameter import (
-    OptimalScalingParameter,
-)
-from pypesto.hierarchical.optimal_scaling.solver import (
+from pypesto.hierarchical.ordinal.parameter import OptimalScalingParameter
+from pypesto.hierarchical.ordinal.solver import (
     compute_interval_constraints,
     get_surrogate_all,
 )
@@ -285,7 +283,7 @@ def _inner_problem_exp():
 
     # Construct inner problem
     inner_problem = OptimalScalingProblem(
-        xs=inner_parameters, data=[data], method=STANDARD
+        xs=inner_parameters, data=[data], edatas=None, method=STANDARD
     )
 
     return inner_problem, expected_inner_parameter_values, simulation
