+++ test/hierarchical/test_spline.py
@@ -16,14 +16,14 @@ from pypesto.C import (
     OPTIMIZE_NOISE,
     InnerParameterType,
 )
-from pypesto.hierarchical.spline_approximation import (
-    SplineInnerProblem,
-    SplineInnerSolver,
+from pypesto.hierarchical.semiquantitative import (
+    SemiquantitativeInnerProblem,
+    SemiquantitativeInnerSolver,
 )
-from pypesto.hierarchical.spline_approximation.parameter import (
-    SplineInnerParameter,
+from pypesto.hierarchical.semiquantitative.parameter import (
+    SemiquantitativeInnerParameter,
 )
-from pypesto.hierarchical.spline_approximation.solver import (
+from pypesto.hierarchical.semiquantitative.solver import (
     _calculate_nllh_for_group,
     _calculate_regularization_for_group,
     _calculate_regularization_gradient_for_group,
@@ -270,7 +270,7 @@ def _inner_problem_exp():
     mask = [np.full(len(simulation), True)]
 
     inner_parameters = [
-        SplineInnerParameter(
+        SemiquantitativeInnerParameter(
             inner_parameter_id=f'{par_type}_{1}_{par_index+1}',
             inner_parameter_type=InnerParameterType.SPLINE,
             scale=LIN,
@@ -284,8 +284,11 @@ def _inner_problem_exp():
         for par_index in range(n_spline_pars)
     ]
 
-    inner_problem = SplineInnerProblem(
-        xs=inner_parameters, data=[data], spline_ratio=spline_ratio
+    inner_problem = SemiquantitativeInnerProblem(
+        xs=inner_parameters,
+        data=[data],
+        edatas=None,
+        spline_ratio=spline_ratio,
     )
 
     return inner_problem, expected_values, simulation, sigma
@@ -310,7 +313,7 @@ def test_spline_inner_solver():
     results = {}
 
     for minimal_diff, option in options.items():
-        inner_solvers[minimal_diff] = SplineInnerSolver(
+        inner_solvers[minimal_diff] = SemiquantitativeInnerSolver(
             options=option,
         )
